wip subclasses indexed
