+++ src/module/heroCreationToolApp.ts
@@ -17,7 +17,7 @@ import { Step } from './step';
 import HeroOption from './options/heroOption';
 import type { ActorDataConstructorData } from '@league-of-foundry-developers/foundry-vtt-types/src/foundry/common/data/data.mjs/actorData';
 import { HitDie, HpCalculation } from './hitDie';
-import { hydrateItems, IndexEntry } from './indexUtils';
+import { IndexEntry, hydrateItems } from './indexes/indexUtils';
 import { LOG_PREFIX, MODULE_ID, MYSTERY_MAN, CLASS_LEVEL } from './constants';
 
 enum StepIndex {
