--- src/module/indexUtils.ts
@@ -1,283 +1,0 @@
-import { DEFAULT_PACKS, LOG_PREFIX, MODULE_ID } from './constants';
-import SettingKeys, { Source, SourceType } from './settings';
-import { getGame } from './utils';
-
-export async function buildSourceIndexes() {
-  console.log(`${LOG_PREFIX} | Indexing source compendiums`);
-  const sourcePacks: Source = (await game.settings.get(MODULE_ID, SettingKeys.SOURCES)) as Source;
-  const itemsPromises: Promise<Item | null | undefined>[] = [];
-  game.packs
-    .filter((p) => p.documentName == 'Item')
-    .forEach((p) => {
-      const name = p.collection;
-      const fieldsToIndex = new Set<string>();
-
-      // name added by default on all when indexed
-      addRaceFields(fieldsToIndex, sourcePacks, name);
-      addRacialFeaturesFields(fieldsToIndex, sourcePacks, name);
-      addClassFields(fieldsToIndex, sourcePacks, name);
-      addClassFeaturesFields(fieldsToIndex, sourcePacks, name);
-      addSpellFields(fieldsToIndex, sourcePacks, name);
-      addFeatFields(fieldsToIndex, sourcePacks, name);
-      addBackgroundFields(fieldsToIndex, sourcePacks, name);
-      addEquipmentFields(fieldsToIndex, sourcePacks, name);
-
-      if (fieldsToIndex.size) {
-        fieldsToIndex.add('img');
-        itemsPromises.push((p as any).getIndex({ fields: [...fieldsToIndex] }));
-      }
-    });
-  await Promise.all(itemsPromises);
-}
-
-export async function buildJournalIndexes() {
-  console.log(`${LOG_PREFIX} | Indexing journals (rules)`);
-  const rulesPack = game.packs.get(DEFAULT_PACKS.RULES);
-  if (!rulesPack) {
-    throw new Error(
-      `${LOG_PREFIX} | Cannot find default DnD5e rules compendium (for indexing sidepanel rules) under name ${DEFAULT_PACKS.RULES}`,
-    );
-  }
-  await (rulesPack as any).getIndex({ fields: ['name', 'content'] });
-}
-
-export async function getRaceEntries() {
-  const raceEntries = await (getIndexEntriesForSource(SourceType.RACES) as unknown as Promise<RaceEntry[]>);
-  // sanitize entries to remove anything nonconforming to a Feature (for now, until Race becomes a type)
-  return raceEntries.filter((r) => r.type == 'feat');
-}
-
-export async function getRaceFeatureEntries() {
-  const raceFeatureEntries = await (getIndexEntriesForSource(SourceType.RACIAL_FEATURES) as unknown as Promise<
-    RacialFeatureEntry[]
-  >);
-  // sanitize entries to remove anything nonconforming to a Feature (for now at least, if Race Features become a type in the future)
-  return raceFeatureEntries.filter((f) => f.type == 'feat' && f?.data?.requirements !== '');
-}
-
-export async function getClassEntries() {
-  const classEntries = await (getIndexEntriesForSource(SourceType.CLASSES) as unknown as Promise<ClassEntry[]>);
-  // sanitize entries to remove anything nonconforming to a Class
-  return classEntries.filter((c) => c.type == 'class');
-}
-
-export async function getClassFeatureEntries() {
-  const classFeatureEntries = await (getIndexEntriesForSource(SourceType.CLASS_FEATURES) as unknown as Promise<
-    RacialFeatureEntry[]
-  >);
-  // sanitize entries to remove anything nonconforming to a Feature (for now at least, if Class Features become a type in the future)
-  return classFeatureEntries.filter((f) => f.type == 'feat' && f?.data?.requirements !== '');
-}
-
-export async function getSpellEntries() {
-  const spellEntries = await (getIndexEntriesForSource(SourceType.SPELLS) as unknown as Promise<SpellEntry[]>);
-  // sanitize entries to remove anything nonconforming to a Spell
-  return spellEntries.filter((s) => s.type == 'spell');
-}
-
-export async function getBackgroundEntries() {
-  const backgroundEntries = await (getIndexEntriesForSource(SourceType.BACKGROUNDS) as unknown as Promise<
-    BackgroundEntry[]
-  >);
-  // sanitize entries to remove anything nonconforming to a Feature (for now at least, if Background Features become a type in the future)
-  return backgroundEntries.filter((f) => f.type == 'background');
-}
-
-export async function getFeatEntries() {
-  const featEntries = await (getIndexEntriesForSource(SourceType.FEATS) as unknown as Promise<FeatEntry[]>);
-  // sanitize entries to remove anything nonconforming to a Feature (for now at least, if Feats become a type in the future)
-  return featEntries.filter((f) => f.type == 'feat');
-}
-
-export async function getEquipmentEntries() {
-  const equipmentEntries = await (getIndexEntriesForSource(SourceType.ITEMS) as unknown as Promise<EquipmentEntry[]>);
-  // sanitize entries to remove anything nonconforming to an Item
-  return equipmentEntries.filter((e) => !['class', 'feat', 'spell'].includes(e.type));
-}
-
-export async function getRuleJournalEntryByName(journalName: string) {
-  const entries = await (game.packs.get(DEFAULT_PACKS.RULES)?.index as unknown as Promise<RuleEntry[]>);
-  return entries.find((e) => e.name === journalName);
-}
-
-async function getIndexEntriesForSource(source: keyof Source) {
-  const sources: Source = (await game.settings.get(MODULE_ID, SettingKeys.SOURCES)) as Source;
-
-  const indexEntries = [];
-  for (const packName of sources[source]) {
-    const pack = game.packs.get(packName);
-    if (!pack) ui.notifications?.warn(`No pack for name [${packName}]!`);
-    if (pack?.documentName !== 'Item') throw new Error(`${packName} is not an Item pack`);
-    const itemPack = pack as CompendiumCollection<CompendiumCollection.Metadata & { entity: 'Item' }>;
-    if ((itemPack as any).indexed) {
-      const packIndexEntries = [...(await itemPack.index)];
-      indexEntries.push(...packIndexEntries.map((e) => ({ ...e, _pack: packName })));
-    } else {
-      console.error(`Index not built for pack [${packName}] - skipping it`);
-    }
-  }
-  return indexEntries;
-}
-
-export async function hydrateItems(indexEntries: Array<IndexEntry>): Promise<Item[]> {
-  console.log(`${LOG_PREFIX} | Hydrating items:`);
-  const worldItems = game.items;
-  if (!worldItems) throw new Error('game.items not initialized yet');
-
-  const itemPromises = indexEntries.map(async (indexEntry) => {
-    if ((indexEntry as any).custom) {
-      return indexEntry; // return custom items as-is
-    }
-    const quantity = (indexEntry as any).data?.quantity;
-    const item = await game.packs.get(indexEntry._pack)?.getDocument(indexEntry._id);
-    if (!item) throw new Error(`No item for id ${indexEntry._id}!`);
-    const itemForEmbedding = worldItems.fromCompendium(item as Item);
-    if (quantity) {
-      (itemForEmbedding!.data as any).quantity = quantity;
-    }
-    return itemForEmbedding;
-  });
-  return (await Promise.all(itemPromises)) as any;
-}
-
-// Entry types & functions that add corresponding fields for index creation
-export type IndexEntry = {
-  _id: string;
-  _pack: string;
-  type: string;
-  name: string;
-  img: string;
-};
-
-export type RaceEntry = IndexEntry & {
-  data: {
-    requirements: string;
-    description: { value: string };
-  };
-};
-function addRaceFields(fieldsToIndex: Set<string>, source: Source, packName: string) {
-  if (source[SourceType.RACES].includes(packName)) {
-    fieldsToIndex.add('data.requirements'); // for figuring subraces
-    fieldsToIndex.add('data.description.value'); // for sidebar
-  }
-}
-
-export type RacialFeatureEntry = IndexEntry & {
-  data: { requirements: string };
-};
-function addRacialFeaturesFields(fieldsToIndex: Set<string>, source: Source, packName: string) {
-  if (source[SourceType.RACIAL_FEATURES].includes(packName)) {
-    fieldsToIndex.add('data.requirements'); // for mapping racial features to races/subraces
-  }
-}
-
-export type ClassEntry = IndexEntry & {
-  data: {
-    description: { value: string };
-    hitDice: string;
-    saves: string[];
-    skills: {
-      number: number;
-      choices: string[];
-      value: string[];
-    };
-    spellcasting: {
-      ability: string;
-      progression: string;
-    };
-    levels: number;
-  };
-};
-function addClassFields(fieldsToIndex: Set<string>, source: Source, packName: string) {
-  if (source[SourceType.CLASSES].includes(packName)) {
-    fieldsToIndex.add('data.description.value'); // for sidebar
-    fieldsToIndex.add('data.hitDice');
-    fieldsToIndex.add('data.saves');
-    fieldsToIndex.add('data.skills');
-    fieldsToIndex.add('data.spellcasting');
-  }
-}
-
-export type ClassFeatureEntry = IndexEntry & {
-  data: {
-    description: { value: string };
-    requirements: string;
-  };
-};
-function addClassFeaturesFields(fieldsToIndex: Set<string>, source: Source, packName: string) {
-  if (source[SourceType.CLASS_FEATURES].includes(packName)) {
-    fieldsToIndex.add('data.requirements'); // for mapping class features to classes
-    fieldsToIndex.add('data.description'); // used to show Spellcasting/Pact Magic features on the Spells tab
-  }
-}
-
-export type SpellEntry = IndexEntry & {
-  data: {
-    level: number;
-  };
-};
-function addSpellFields(fieldsToIndex: Set<string>, source: Source, packName: string) {
-  if (source[SourceType.SPELLS].includes(packName)) {
-    fieldsToIndex.add('data.level');
-  }
-}
-
-export type FeatEntry = IndexEntry & {
-  data: { requirements: string };
-};
-function addFeatFields(fieldsToIndex: Set<string>, source: Source, packName: string) {
-  if (source[SourceType.FEATS].includes(packName)) {
-    fieldsToIndex.add('data.requirements'); // TODO if feat has a requirement show it.
-  }
-}
-
-export type BackgroundEntry = IndexEntry & unknown;
-function addBackgroundFields(fieldsToIndex: Set<string>, source: Source, packName: string) {
-  if (source[SourceType.BACKGROUNDS].includes(packName)) {
-    fieldsToIndex.add('name');
-  }
-}
-
-export type EquipmentEntry = IndexEntry & {
-  data: {
-    price: number;
-    rarity: string;
-    quantity?: number;
-  };
-};
-function addEquipmentFields(fieldsToIndex: Set<string>, source: Source, packName: string) {
-  if (source[SourceType.ITEMS].includes(packName)) {
-    fieldsToIndex.add('data.price');
-    fieldsToIndex.add('data.rarity');
-    fieldsToIndex.add('data.quantity');
-    //fieldsToIndex.add('data.description'); maybe description to find Spellcasting Foci ?
-  }
-}
-
-export type RuleEntry = {
-  _id: string;
-  name: string;
-  content: string;
-};
-
-export function getIndexByUuid(uuid: string): IndexEntry | undefined {
-  const { pack, id } = parseUuid(uuid);
-
-  // TODO see if its correctly indexed
-
-  return {
-    ...(getGame()
-      .packs.get(pack)
-      ?.index.find((i) => i._id === id) as IndexEntry),
-    _pack: pack,
-  };
-}
-
-function parseUuid(uuid: string): { pack: any; id: any } {
-  const firstDot = uuid.indexOf('.');
-  const lastDot = uuid.lastIndexOf('.');
-
-  const pack = uuid.substring(firstDot + 1, lastDot);
-  const id = uuid.substring(lastDot + 1);
-  return { pack, id };
-}
