+++ src/module/options/fixedOption.ts
@@ -1,7 +1,7 @@
 import { StepEnum } from '../step';
 import { ActorDataConstructorData } from '@league-of-foundry-developers/foundry-vtt-types/src/foundry/common/data/data.mjs/actorData';
 import HeroOption, { apply } from './heroOption';
-import { IndexEntry } from '../indexUtils';
+import { IndexEntry } from '../indexes/indexUtils';
 
 /**
  * Represents a fixed value that will be imprinted into the created actor
