+++ src/module/options/quantifiableOption.ts
@@ -1,7 +1,7 @@
 import { StepEnum } from '../step';
 import { ActorDataConstructorData } from '@league-of-foundry-developers/foundry-vtt-types/src/foundry/common/data/data.mjs/actorData';
 import HeroOption, { apply } from './heroOption';
-import { EquipmentEntry } from '../indexUtils';
+import { EquipmentEntry } from '../indexes/indexUtils';
 
 /**
  * Represents a value with a selectable quantity that will be imprinted into the created actor
