+++ src/module/options/searchableIndexEntryOption.ts
@@ -1,7 +1,7 @@
 import { StepEnum } from '../step';
 import { ActorDataConstructorData } from '@league-of-foundry-developers/foundry-vtt-types/src/foundry/common/data/data.mjs/actorData';
 import HeroOption, { apply } from './heroOption';
-import { IndexEntry } from '../indexUtils';
+import { IndexEntry } from '../indexes/indexUtils';
 import { MYSTERY_MAN } from '../constants';
 
 /**
