+++ src/templates/tabs/background.html
@@ -13,13 +13,14 @@
                 </div>
                 <div data-hct_area="feature">
                     <h2 class="hct-mb-md hct-font-title">{{localize 'HCT.Common.Feature'}}</h2>
-                    <div class="hct-w-full flexrow hct-justify-between hct-pb-sm">
-                        <span>{{ localize 'HCT.Background.ShowFeatureOnSide'}}</span>
-                        <input type="checkbox" data-hct-show-background-feature-desc>
-                    </div>
                 </div>
             </div>
         </div>
-        <div class="hct-panel-right hct-infoArea" data-hct_background_description></div>
+        <div class="hct-panel-right hct-infoArea" data-hct_background_description>
+            <div class="hct-w-full flexrow hct-justify-between hct-pb-sm">
+                <span>{{ localize 'HCT.Background.ShowFeatureOnSide'}}</span>
+                <input type="checkbox" data-hct-show-background-feature-desc>
+            </div>
+        </div>
     </main>
 </section>
